# feat(benchmark): add comprehensive benchmark infrastructure