--- a/.github/workflows/benchmarks.yml
+++ b/.github/workflows/benchmarks.yml
@@ -68,7 +68,7 @@ jobs:
 
     - name: Run benchmarks
       run: |
-        cd build/benchmarks
+        cd build/bin
         ./database_benchmarks \
           --benchmark_format=json \
           --benchmark_out=benchmark_results_${{ matrix.os }}.json \
@@ -77,7 +77,7 @@ jobs:
 
     - name: Run benchmarks (console output)
       run: |
-        cd build/benchmarks
+        cd build/bin
         ./database_benchmarks \
           --benchmark_filter="QueryBuilder_Create|ConnectionPool_Acquire|Transaction_Begin" \
           --benchmark_repetitions=3
@@ -86,14 +86,14 @@ jobs:
       uses: actions/upload-artifact@v4
       with:
         name: benchmark-results-${{ matrix.os }}
-        path: build/benchmarks/benchmark_results_${{ matrix.os }}.json
+        path: build/bin/benchmark_results_${{ matrix.os }}.json
         retention-days: 30
 
     - name: Save baseline (if requested)
       if: github.event.inputs.save_baseline == 'true' && github.ref == 'refs/heads/main'
       run: |
         mkdir -p benchmarks/baselines
-        cp build/benchmarks/benchmark_results_${{ matrix.os }}.json \
+        cp build/bin/benchmark_results_${{ matrix.os }}.json \
            benchmarks/baselines/baseline_${{ matrix.os }}_$(date +%Y%m%d).json
 
     - name: Compare with baseline (if exists)
@@ -108,7 +108,7 @@ jobs:
 
         if [ -f "$BASELINE" ]; then
           echo "Baseline: $BASELINE"
-          echo "Current: build/benchmarks/benchmark_results_${{ matrix.os }}.json"
+          echo "Current: build/bin/benchmark_results_${{ matrix.os }}.json"
 
           # Download compare.py
           curl -O https://raw.githubusercontent.com/google/benchmark/main/tools/compare.py
@@ -117,7 +117,7 @@ jobs:
           # Run comparison
           python3 compare.py \
             "$BASELINE" \
-            build/benchmarks/benchmark_results_${{ matrix.os }}.json
+            build/bin/benchmark_results_${{ matrix.os }}.json
         fi
 
     - name: Check for performance regression