--- a/benchmarks/query_execution_bench.cpp
+++ b/benchmarks/query_execution_bench.cpp
@@ -6,15 +6,15 @@
 
 #include <benchmark/benchmark.h>
 #include "database/query_builder.h"
-#include "database/database_manager.h"
+#include "database/database_types.h"
 #include <memory>
 
-using namespace database_module;
+using namespace database;
 
 // Benchmark query builder creation
 static void BM_QueryBuilder_Create(benchmark::State& state) {
     for (auto _ : state) {
-        query_builder qb;
+        query_builder qb(database_types::postgres);
         benchmark::DoNotOptimize(qb);
     }
 }
@@ -23,7 +23,7 @@ BENCHMARK(BM_QueryBuilder_Create);
 // Benchmark SELECT query building
 static void BM_QueryBuilder_SelectSimple(benchmark::State& state) {
     for (auto _ : state) {
-        query_builder qb;
+        query_builder qb(database_types::postgres);
         auto query = qb.select({"id", "name", "email"})
                        .from("users")
                        .build();
@@ -35,10 +35,10 @@ BENCHMARK(BM_QueryBuilder_SelectSimple);
 // Benchmark SELECT with WHERE clause
 static void BM_QueryBuilder_SelectWithWhere(benchmark::State& state) {
     for (auto _ : state) {
-        query_builder qb;
+        query_builder qb(database_types::postgres);
         auto query = qb.select({"id", "name", "email"})
                        .from("users")
-                       .where("age > 18")
+                       .where("age", ">", 18)
                        .build();
         benchmark::DoNotOptimize(query);
     }
@@ -48,12 +48,12 @@ BENCHMARK(BM_QueryBuilder_SelectWithWhere);
 // Benchmark SELECT with multiple conditions
 static void BM_QueryBuilder_SelectComplex(benchmark::State& state) {
     for (auto _ : state) {
-        query_builder qb;
+        query_builder qb(database_types::postgres);
         auto query = qb.select({"id", "name", "email", "age", "created_at"})
                        .from("users")
-                       .where("age > 18")
-                       .where("status = 'active'")
-                       .order_by("created_at DESC")
+                       .where("age", ">", 18)
+                       .where("status", "=", "active")
+                       .order_by("created_at", sort_order::desc)
                        .limit(100)
                        .build();
         benchmark::DoNotOptimize(query);
@@ -64,10 +64,12 @@ BENCHMARK(BM_QueryBuilder_SelectComplex);
 // Benchmark INSERT query building
 static void BM_QueryBuilder_Insert(benchmark::State& state) {
     for (auto _ : state) {
-        query_builder qb;
-        auto query = qb.insert_into("users")
-                       .values({{"name", "John"}, {"email", "john@example.com"}})
-                       .build();
+        query_builder qb(database_types::postgres);
+        std::map<std::string, database_value> data = {
+            {"name", "John"},
+            {"email", "john@example.com"}
+        };
+        auto query = qb.insert(data).build();
         benchmark::DoNotOptimize(query);
     }
 }
@@ -76,10 +78,13 @@ BENCHMARK(BM_QueryBuilder_Insert);
 // Benchmark UPDATE query building
 static void BM_QueryBuilder_Update(benchmark::State& state) {
     for (auto _ : state) {
-        query_builder qb;
-        auto query = qb.update("users")
-                       .set({{"name", "Jane"}, {"email", "jane@example.com"}})
-                       .where("id = 123")
+        query_builder qb(database_types::postgres);
+        std::map<std::string, database_value> data = {
+            {"name", "Jane"},
+            {"email", "jane@example.com"}
+        };
+        auto query = qb.update(data)
+                       .where("id", "=", 123)
                        .build();
         benchmark::DoNotOptimize(query);
     }
@@ -89,9 +94,9 @@ BENCHMARK(BM_QueryBuilder_Update);
 // Benchmark DELETE query building
 static void BM_QueryBuilder_Delete(benchmark::State& state) {
     for (auto _ : state) {
-        query_builder qb;
-        auto query = qb.delete_from("users")
-                       .where("id = 123")
+        query_builder qb(database_types::postgres);
+        auto query = qb.remove()
+                       .where("id", "=", 123)
                        .build();
         benchmark::DoNotOptimize(query);
     }
@@ -101,11 +106,11 @@ BENCHMARK(BM_QueryBuilder_Delete);
 // Benchmark JOIN query building
 static void BM_QueryBuilder_Join(benchmark::State& state) {
     for (auto _ : state) {
-        query_builder qb;
+        query_builder qb(database_types::postgres);
         auto query = qb.select({"u.id", "u.name", "o.order_id", "o.total"})
                        .from("users u")
                        .join("orders o", "u.id = o.user_id")
-                       .where("o.status = 'completed'")
+                       .where("o.status", "=", "completed")
                        .build();
         benchmark::DoNotOptimize(query);
     }
@@ -115,10 +120,10 @@ BENCHMARK(BM_QueryBuilder_Join);
 // Benchmark parameterized query building
 static void BM_QueryBuilder_Parameterized(benchmark::State& state) {
     for (auto _ : state) {
-        query_builder qb;
+        query_builder qb(database_types::postgres);
         auto query = qb.select({"*"})
                        .from("users")
-                       .where("email = ?")
+                       .where("email", "=", "user@example.com")
                        .build();
         benchmark::DoNotOptimize(query);
     }
@@ -134,7 +139,7 @@ static void BM_QueryBuilder_ComplexityScaling(benchmark::State& state) {
     }
 
     for (auto _ : state) {
-        query_builder qb;
+        query_builder qb(database_types::postgres);
         auto query = qb.select(columns)
                        .from("test_table")
                        .build();